backend-django/requirements.txt
Django>=4.2
djangorestframework>=3.14
PyJWT>=2.8
dj-database-url>=1.0
python-dotenv>=1.0
gunicorn>=20.1